GP EditMode: Simplify the way that this is handled
Now, instead of actually altering the ob->mode flag, we just make the
object mode setter operator (and the header) delegate to the appropriate
GPencil things to do their thing if they wish first.

As a result, we don't have to worry about version patching or issues with
changing selected/active objects.