Stop-Gap fix: Proportional editing was broken in new files for GPencil
Looks like my earlier commit wasn't quite right. For now, just reverting these
so that I can use this build for Siggraph; will look into a better fix later.

(The problem seems to be that the gp flag is getting auto-set too late - needs
to be done in initTransInfo() for the prop edit to take hold)