Rework workflow for connecting Face Maps and bones
Previously, a Face Map was 'connected to a bone' by giving it the same name as the Face Map, and assigning the armature which includes the bone to the object's Face Map using an Armature modifier. This was pretty unintuitive and tiresome to set up and it also brought some technical issues (there was no nice way to find a Face Map assigned to a bone from the bone - this means we couldn't use Face Maps together with bones in pose mode as current design approach would require).

So finally I came up with a new approach:
The bone context in the Properties Editor got a new panel "Face Maps". In this panel there is a search button to choose an object. After an object is chosen, a Face Map from this object can be chosen using a second search button. That means a user can assign a Face Map to a bone directly, without having to worry about the armature.

This approach currently has the limitation that only one Face Map can be assigned to a bone (a Face Map can be used by multiple bones though). This means we can later use one Face Map to manipulate multiple bones, but we can't use multiple Face Maps to manipulate one bone.
Anyway, this could be changed easily later, but I'd rather avoid this since it complicates the UI/workflow again.

----

Old code is kept.
Also includes minor cleanup.