GP Editing: Snapping Tools (Shift S)
This commit introduces the ability to snap GP points to the cursor/grid,
as well as snapping the cursor to GP points. Currently this functionality
is only available in the 3D view (since the cursor/grid settings are only
available in that view), where they are most needed.

Later, some of these operators can be extended to other editors on an
"as-needed" basis, based on finding some way of dealing with the lack
of 2D cursor in those views.

Coded during my recent travels :)