GPencil Dopesheet: Filtering option to only include GPencil layers related to the active scene
As requested by the Caminandes team, in the "Grease Pencil" mode of the dopesheet,
it is now possible to view only the Grease Pencil blocks+layers which come from the
active scene and/or the objects in that scene. This is useful when working with
multiple scenes, or when there are many other GPencil datablocks in the file for
various editors that aren't visible.

To use, enable the 'Active Only' toggle with the scene icon (located beside the
Summary toggle). By default, this is not enabled, so that backwards comptability
with old files is not affected (i.e. users are less likely to why they are no longer
able to see keyframes that would've previously been easily visible).