GPencil: Moved Stroke Placement settings from GP datablock to tool settings
This commit moves the Stroke Placement settings from being a per-datablock
setting to being stored in tool settings instead. The main reason this was
done was to resolve a usability issue with the old way: in order to set this
setting, you had to first create a GP datablock to be able to see the settings!

Now, by having these as tool settings (as they should have been from the start),
they can be always visible (even when there isn't a datablock yet), meaning that
it will work as expected from the very first stroke.


Implementation Notes:
* There are separate Stroke Placement properties for the different editor types.
  e.g. 3D View, Sequencer Preview, Image Editor, and all other 2D Views

* I couldn't get the version patching code working (to port over placement settings
  from the GP datablock attached to the active scene), as this was causing crashes
  and other general weirdness when loading old files. The code to do this is there,
  but #if 0'ed out.