GPencil Drawing Usability: "Additive Drawing"
This commit adds a new option which will make it easier to use the Grease Pencil
drawing tools for animation. With the "Additive Drawing" option enabled in the
toolshelf (located alongside the "Continuous Drawing" toggle), the active frame's
strokes will be carried over/copied if you start drawing on an empty frame (i.e.
one without any keyframe already). This saves the effort of keeping a dopesshet
open, and to remember to duplicate the current frame before starting to draw the
next pose (or risk managing to draw the perfect pose, but without everything else).

Examples of cases where this comes in handy includes animating facial expressions
(when all outlines are on the same layer), or animating "growing" things (e.g. vines,
or concentric circles growing from a central point).

Note: Even without this option enabled, this is the default behaviour when using
the eraser on an "empty" frame. This makes it easier to do shots where you're just
changing parts of the facial expression, or if you're animating an "eraser" effect.

Note 2: The naming of this feature could do with some work. I struggled with several
alternative names for it, but this one seems the most useful about what it helps users
do. Suggestions welcome!