Initial steps for porting wmWidget types to C++
Antony and myself came to the conclusion that widgets could make good use of C++ features, so I decided to try porting them over.
Main changes in this commit are:
* Introduce object oriented file structure
* Port some wmWidget structs to classes
* Add initial C API
* Actually use some of this stuff

Some things are a bit tricky to port over (ListBase, registering widgets in rna_ files for Python API, etc), but they're definitely solvable.

NOTE: This branch is completely unstable for now, but that is expected. Also some parts of the Python API are commented out now (I'd be surprised if it still worked anyway :S).