GPencil Eraser: Eraser is now pressure sensitive
Instead of immediately erasing points, the eraser now first reduces the thickness
of the stroke points before culling the ones that have become overly thin (and
cannot be rendered properly anyway as a result). It is also pressure sensitive now,
and has a linear falloff effect (i.e. points further away from the center of the
eraser circle get affected less). This should make it easier to use to eraser to
make fine adjustments to your sketches.

Notes:
* The eraser has been rewritten to use a new algorithm. Specifically, it now takes
  advantage of the stroke point deletion methods developed for the Delete operator
  instead of it's own buggy split-and-delete methods (which were only removing
  a single point from each stroke each time).

* In tests so far, it seems to allow a wider range of control over what gets erased
  and what doesn't. Pressing hard (or using a mouse) should still just erase everything
  with relative ease still.

* The way that this behaves could still do with some tweaking. In particular, care
  is needed when trying to use the eraser to "lighten up" tips of thin bunches of
  strokes (e.g. eyebrows/moustaches), as it still tends to gobble these up too much.
  We could probably do with some tool settings to control the eraser strength.