Pushes a first-pass revision to the original OpenColorIO configurations.
This removes legacy and incorrect transformations, as well
as establishes a cleaner family organization for UI
interactions. It also adds AdobeRGB 1999 compatible image
transforms, which would be of use to texturers. In addition
to this, the DCI-P3 display group has a single entity which
permits late model 2015 iMacs to properly view Blender
output. Finally, it adds the "chromaticity" role, which is
of importance to colourspace agnostic rendering and image
manipulations.