Cycles: Add optional Blue-Noise Dithered Sobol Sampling
This patch implements precomputed dithered sampling as described in the paper "Blue-noise Dithered Sampling".

Mainly for low sample counts, such as the first few seconds of viewport rendering, the improvement in quality is substantial - especially the brightness flickering happening with regular Sobol and HDRIs is pretty much gone.

For now, the dithering matrix included in sobol.cpp isn't perfect yet - I'm currently running a 400 million iterations simulated annealing process, but that will still take about a day to finish. I'll update the patch as soon as it's here.
Because of the temporary matrix, you can still see some tiling happening at low sample counts, but with a good dithering matrix that shouldn't be as visible.

Reviewers: #cycles

Differential Revision: https://developer.blender.org/D2149

Conflicts:
	intern/cycles/kernel/kernel_types.h
	intern/cycles/util/util_texture.h

Conflicts:
	intern/cycles/kernel/kernel_types.h

Conflicts:
	intern/cycles/kernel/kernel_random.h
	intern/cycles/kernel/kernel_types.h
	intern/cycles/kernel/kernel_volume.h