Cycles: Refactor OpenCL cache to be able to cache split kernel programs
Previously, the OpenCL cache would store the context and Base- and Megakernel for every device. With this change, it stores context and a ustring->cl_program map that can hold arbitrary programs.
The old two kernels are stored with constant keys, but now it can also be used to store split kernel programs, where the key consists of the name and the device md5 (no need to include the source md5 here since the sources shouldn't change during runtime).

Depends on D2231.

The main downside is a bit of duplicated code in the Cache functions, but there are enough differences to make a generic function basically impossible (unless you use tons of template magic, which ends up longer than the duplicated code).

Reviewers: sergey, brecht, dingto, juicyfruit

Differential Revision: https://developer.blender.org/D2233