Cycles: Optimize shaders earlier to skip unneccessary attributes for noninteractive rendering
Before, Cycles would first sync the shader exactly as built, then determine and sync the used attributes and then optimize the shader.
Therefore, even completely unconnected nodes could cause unneccessary attributes to be synced.

The reason for this is to avoid frequent resyncs when editing shaders interactively, but it can still be avoided for final renders - which is what this commit does.

Reviewers:

Subscribers: