Add SOCK_VIRTUAL for better indication of Render Result node sockets
To provide better visual indication in the compositor, sockets for passes that aren't availiable are still drawn, but all their links are dashed and they behave as if SOCK_UNAVAIL was set. This is handled by the new socket flag SOCK_VIRTUAL.
This is to avoid confusion since, when the user hasn't rendered after loading a file, only the Image and Alpha sockets are available. Therefore, without this change, the compositor setup would appear broken until the first render.

Conflicts:
	source/blender/editors/space_node/drawnode.c